Move system-probe module API into module subdir
