+++ cmd/system-probe/api/module/common.go
@@ -1,4 +1,4 @@
-package api
+package module
 
 import (
 	"errors"
