+++ cmd/system-probe/api/module/loader.go
@@ -1,4 +1,4 @@
-package api
+package module
 
 import (
 	"net/http"
