+++ cmd/system-probe/modules/all_linux.go
@@ -2,12 +2,10 @@
 
 package modules
 
-import (
-	"github.com/DataDog/datadog-agent/cmd/system-probe/api"
-)
+import "github.com/DataDog/datadog-agent/cmd/system-probe/api/module"
 
 // All System Probe modules should register their factories here
-var All = []api.Factory{
+var All = []module.Factory{
 	NetworkTracer,
 	TCPQueueLength,
 	OOMKillProbe,
