+++ cmd/system-probe/modules/all_unsupported.go
@@ -2,7 +2,7 @@
 
 package modules
 
-import "github.com/DataDog/datadog-agent/cmd/system-probe/api"
+import "github.com/DataDog/datadog-agent/cmd/system-probe/api/module"
 
 // All System Probe modules should register their factories here
-var All = []api.Factory{}
+var All = []module.Factory{}
