+++ cmd/system-probe/modules/all_windows.go
@@ -2,11 +2,9 @@
 
 package modules
 
-import (
-	"github.com/DataDog/datadog-agent/cmd/system-probe/api"
-)
+import "github.com/DataDog/datadog-agent/cmd/system-probe/api/module"
 
 // All System Probe modules should register their factories here
-var All = []api.Factory{
+var All = []module.Factory{
 	NetworkTracer,
 }
