+++ cmd/system-probe/modules/oom_kill_probe.go
@@ -5,7 +5,7 @@ package modules
 import (
 	"net/http"
 
-	"github.com/DataDog/datadog-agent/cmd/system-probe/api"
+	"github.com/DataDog/datadog-agent/cmd/system-probe/api/module"
 	"github.com/DataDog/datadog-agent/cmd/system-probe/config"
 	"github.com/DataDog/datadog-agent/cmd/system-probe/utils"
 	"github.com/DataDog/datadog-agent/pkg/collector/corechecks/ebpf/probe"
@@ -15,9 +15,9 @@ import (
 )
 
 // OOMKillProbe Factory
-var OOMKillProbe = api.Factory{
+var OOMKillProbe = module.Factory{
 	Name: config.OOMKillProbeModule,
-	Fn: func(cfg *config.Config) (api.Module, error) {
+	Fn: func(cfg *config.Config) (module.Module, error) {
 		log.Infof("Starting the OOM Kill probe")
 		okp, err := probe.NewOOMKillProbe(ebpf.NewConfig())
 		if err != nil {
@@ -27,7 +27,7 @@ var OOMKillProbe = api.Factory{
 	},
 }
 
-var _ api.Module = &oomKillModule{}
+var _ module.Module = &oomKillModule{}
 
 type oomKillModule struct {
 	*probe.OOMKillProbe
