+++ cmd/system-probe/modules/process.go
@@ -10,7 +10,7 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/DataDog/datadog-agent/cmd/system-probe/api"
+	"github.com/DataDog/datadog-agent/cmd/system-probe/api/module"
 	"github.com/DataDog/datadog-agent/cmd/system-probe/config"
 	"github.com/DataDog/datadog-agent/pkg/process/encoding"
 	"github.com/DataDog/datadog-agent/pkg/process/procutil"
@@ -21,9 +21,9 @@ import (
 var ErrProcessUnsupported = errors.New("process module unsupported")
 
 // Process is a module that fetches process level data
-var Process = api.Factory{
+var Process = module.Factory{
 	Name: config.ProcessModule,
-	Fn: func(cfg *config.Config) (api.Module, error) {
+	Fn: func(cfg *config.Config) (module.Module, error) {
 		log.Infof("Creating process module for: %s", filepath.Base(os.Args[0]))
 
 		// we disable returning zero values for stats to reduce parsing work on process-agent side
@@ -35,7 +35,7 @@ var Process = api.Factory{
 	},
 }
 
-var _ api.Module = &process{}
+var _ module.Module = &process{}
 
 type process struct{ probe *procutil.Probe }
