+++ pkg/security/module/module_unsupported.go
@@ -8,12 +8,12 @@
 package module
 
 import (
-	"github.com/DataDog/datadog-agent/cmd/system-probe/api"
+	"github.com/DataDog/datadog-agent/cmd/system-probe/api/module"
 	"github.com/DataDog/datadog-agent/pkg/ebpf"
 	"github.com/DataDog/datadog-agent/pkg/security/config"
 )
 
 // NewModule instantiates a runtime security system-probe module
-func NewModule(cfg *config.Config) (api.Module, error) {
+func NewModule(cfg *config.Config) (module.Module, error) {
 	return nil, ebpf.ErrNotImplemented
 }
